requirements.txt
asyncio==3.4.3
blinker==1.9.0
click==8.1.7
Flask==3.1.0
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pika==1.3.2
psycopg2==2.9.10
python-dotenv==1.0.1
websockets==13.1
Werkzeug==3.1.3
